[update] APIへのリクエストロジックを別ファイルに移動
